Sync with pdpmake
One big reason to make this fork of maramake was to allow '/' in
target names.  This is now enabled upstream, since there is a
draft 202X Posix standard which now allows '/' and '-' in 'make'
targets.

That in mind, the main reasons to keep this version are:

* Support for '--version'
* Make default CC value 'cc' (breaking POSIX, but informal standard
  among make implementations)
* No GPL code in tree

That’s enough to keep this around.